Refactor out WikiForumGUI functions around category and forum headers
Removes showMainHeader, showListTagHeader, showMainFooter, showListTagFooter

Functions which only return part of the tag are hard to reason about,
making security review difficult if everything returns a different
fragment of HTML.  Ultimately I think a seperate file for GUI was a bad
choice here since HTML is still generated in every file, which means
a display table might have HTML generated in three different files,
with some non-trivial functions.  Reducing the function calls, at the
very least, reduces some of the cognitive load when reading.

Also the CSS remains a mess, overspecified to some deal to reusing
class names, but also to workaround limits in CSS 2.1, which is no
longer relevant in today's browsers.  Fixing a few of those in the
headers.

Change-Id: I3bd5323e7c62c412c0040920b5020ed1e7f11335