Pass $groups (the groups the $user is in) to the WikiForumSig hook + slight cleanup
In the past, it made sense to only pass in the $user because getGroups/getEffectiveGroups were User class methods.
Now that is no longer the case. It seems only logical that any and all subscribers to this hook would also need access to $groups,
so there's no real reason why they should be forced to essentially recalculate that data again when we can just pass the existing value
to the hook.

Change-Id: Icf5b045b9522db54a86a72890df07119443ec943